+++ Source/camera/camera.h
@@ -4,7 +4,7 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
 
-#include "camera_property.h"
+#include "camera_properties.h"
 #include "../rendering/shader.h"
 #include "../window.h"
