+++ Source/camera/camera_properties.h
