--- Source/context.cpp
@@ -1,11 +1,0 @@
-#include "context.h"
-
-
-
-void setupContext(GLFWwindow* window) {
-    glfwSetKeyCallback(window, inputContext);
-}
-
-void inputContext(GLFWwindow* iWindow, int key, int scancode, int action, int mods) {
-    PlayerCallback::processInput(key, action);
-}
