+++ Source/rendering/graphics.cpp
@@ -1,4 +1,4 @@
-#include "graphic.h"
+#include "graphics.h"
 
 
 //VERTEX OBJECT
