+++ Source/window.cpp
@@ -1,18 +1,24 @@
 #include "window.h"
 
 
+GLFWwindow* windowCreation(std::string str_name) {
+	const char* name = str_name.c_str();
+	GLFWwindow* window = glfwCreateWindow(SCR_WIDTH, SCR_HEIGHT, name, NULL, NULL);
 
-void windowLoader(GLFWwindow* window) {
+    //Window verifier
+    if (window == NULL) {
+        std::cerr << "Failed to create GLFW window" << std::endl;
+        glfwTerminate();
+    }
+
+
+	glfwSwapInterval(1);
+
 	glfwMakeContextCurrent(window);
 	glfwSetFramebufferSizeCallback(window, framebufferSizeCallback);
-}
+	glfwSetKeyCallback(window, PlayerCallback::processInput);
 
-int windowVerifier(GLFWwindow* window) {
-	if (window == NULL) {
-		std::cerr << "Failed to create GLFW window" << std::endl;
-		glfwTerminate();
-		return -1;
-	}
+	return window;
 }
 
 //[GLFW] Whenever the window is resized (by OS or user resize) this callback function executes
