+++ Source/window.h
@@ -4,10 +4,11 @@
 
 #include <GLFW/glfw3.h>
 
+#include "input/callbacks.h"
+
 //Screen size settings
 const unsigned int SCR_WIDTH = 800;
 const unsigned int SCR_HEIGHT = 600;
 
-void windowLoader(GLFWwindow* window);
-int windowVerifier(GLFWwindow* window);
+GLFWwindow* windowCreation(std::string str_name);
 void framebufferSizeCallback(GLFWwindow* window, int width, int height);
